[CLEAN] CommandEffectとMessageEffectsをAllSyntaxに移す
