+++ src/main/scala/com/github/unchama/buildassist/menu/BlockPlacementSkillMenu.scala
@@ -7,7 +7,6 @@ import com.github.unchama.menuinventory.slot.button.action.LeftClickButtonEffect
 import com.github.unchama.menuinventory.slot.button.{Button, RecomputedButton}
 import com.github.unchama.menuinventory.{Menu, MenuFrame, MenuSlotLayout}
 import com.github.unchama.seichiassist.CommonSoundEffects
-import com.github.unchama.targetedeffect.MessageEffects._
 import com.github.unchama.targetedeffect.player.FocusedSoundEffect
 import com.github.unchama.{menuinventory, targetedeffect}
 import org.bukkit.ChatColor._
@@ -18,6 +17,7 @@ object BlockPlacementSkillMenu extends Menu {
 
   import com.github.unchama.seichiassist.concurrent.PluginExecutionContexts.layoutPreparationContext
   import com.github.unchama.targetedeffect._
+  import com.github.unchama.targetedeffect.syntax._
   import menuinventory.syntax._
 
   private implicit class PlayerDataOps(val playerData: PlayerData) extends AnyVal {
