+++ src/main/scala/com/github/unchama/contextualexecutor/builder/ArgumentParserScope.scala
@@ -2,8 +2,8 @@ package com.github.unchama.contextualexecutor.builder
 
 import com.github.unchama
 import com.github.unchama.contextualexecutor.builder.TypeAliases.ResponseEffectOrResult
-import com.github.unchama.targetedeffect.MessageEffects._
-import com.github.unchama.targetedeffect.TargetedEffect
+import com.github.unchama.targetedeffect._
+import com.github.unchama.targetedeffect.syntax._
 import org.bukkit.command.CommandSender
 
 /**
