+++ src/main/scala/com/github/unchama/contextualexecutor/executors/PrintUsageExecutor.scala
@@ -2,7 +2,7 @@ package com.github.unchama.contextualexecutor.executors
 
 import cats.effect.IO
 import com.github.unchama.contextualexecutor.{ContextualExecutor, RawCommandContext}
-import com.github.unchama.targetedeffect.MessageEffects._
+import com.github.unchama.targetedeffect.syntax._
 
 /**
  * コマンドの[Command.usage]を送信者に送り返すだけのアクションを返すExecutor
