+++ src/main/scala/com/github/unchama/seichiassist/commands/AchievementCommand.scala
@@ -4,8 +4,8 @@ import cats.effect.IO
 import com.github.unchama.contextualexecutor.builder.{ContextualExecutorBuilder, Parsers}
 import com.github.unchama.contextualexecutor.executors.EchoExecutor
 import com.github.unchama.seichiassist.SeichiAssist
-import com.github.unchama.targetedeffect.MessageEffects._
 import com.github.unchama.targetedeffect.TargetedEffect
+import com.github.unchama.targetedeffect.syntax._
 import org.bukkit.Bukkit
 import org.bukkit.ChatColor._
 import org.bukkit.command.{CommandSender, TabExecutor}
