+++ src/main/scala/com/github/unchama/seichiassist/commands/EffectCommand.scala
@@ -4,7 +4,7 @@ import cats.effect.IO
 import com.github.unchama.contextualexecutor.executors.BranchedExecutor
 import com.github.unchama.seichiassist.SeichiAssist
 import com.github.unchama.seichiassist.commands.contextual.builder.BuilderTemplates.playerCommandBuilder
-import com.github.unchama.targetedeffect.MessageEffects._
+import com.github.unchama.targetedeffect.syntax._
 import com.github.unchama.targetedeffect.{TargetedEffect, emptyEffect}
 import org.bukkit.ChatColor._
 import org.bukkit.command.TabExecutor
