+++ src/main/scala/com/github/unchama/seichiassist/commands/SeichiHasteCommand.scala
@@ -6,8 +6,8 @@ import com.github.unchama.contextualexecutor.executors.EchoExecutor
 import com.github.unchama.seichiassist.SeichiAssist
 import com.github.unchama.seichiassist.data.potioneffect.FastDiggingEffect
 import com.github.unchama.seichiassist.util.TypeConverter
-import com.github.unchama.targetedeffect.MessageEffects._
 import com.github.unchama.targetedeffect.TargetedEffect
+import com.github.unchama.targetedeffect.syntax._
 import enumeratum._
 import org.bukkit.Bukkit
 import org.bukkit.ChatColor._
