+++ src/main/scala/com/github/unchama/seichiassist/commands/SubHomeCommand.scala
@@ -8,7 +8,7 @@ import com.github.unchama.contextualexecutor.builder.Parsers
 import com.github.unchama.contextualexecutor.executors.{BranchedExecutor, EchoExecutor}
 import com.github.unchama.seichiassist.SeichiAssist
 import com.github.unchama.seichiassist.commands.contextual.builder.BuilderTemplates.playerCommandBuilder
-import com.github.unchama.targetedeffect.MessageEffects._
+import com.github.unchama.targetedeffect.syntax._
 import org.bukkit.ChatColor._
 import org.bukkit.command.TabExecutor
