+++ src/main/scala/com/github/unchama/seichiassist/menus/minestack/MineStackButtons.scala
@@ -50,8 +50,8 @@ private[minestack] case class MineStackButtons(player: Player) {
   import MineStackButtons._
   import MineStackObjectCategory._
   import com.github.unchama.seichiassist.concurrent.PluginExecutionContexts.layoutPreparationContext
-  import com.github.unchama.targetedeffect.MessageEffects._
   import com.github.unchama.targetedeffect._
+  import com.github.unchama.targetedeffect.syntax._
   import player._
 
   import scala.jdk.CollectionConverters._
